+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Add remove button to add mods dialog when mod is already installed
 - When adding mods to an instance, get metadata from all mod platforms
 - Add ability to change description for servers
+- Add more info when cannot run server script
 
 ### Fixes
 - Issue with updating/reinstalling non solder Technic packs not showing correct version